# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  server.apiErrorResponse:
    properties:
      error:
        type: string
    type: object
  server.apiResponse:
    properties:
      data: {}
      meta: {}
    type: object
...
  server.platformMFAConfirmRequest:
    properties:
      code:
        type: string
      password:
        type: string
      secret:
        type: string
    required:
    - code
    - password
    - secret
    type: object
  server.platformMFAEnrollRequest:
    properties:
      code:
        type: string
      enrollment_token:
        type: string
      secret:
        type: string
    required:
    - code
    - enrollment_token
    - secret
    type: object
...
info:
  contact: {}
  description: 'Backend service for the NEXQIA Healthcare Operating Platform. All
    routes are under /api/v1. Response envelope: {"data": ..., "meta": {...}}.'
  title: NEXQIA API
  version: "1.0"
paths:
...
  /platform/admin-users/{id}/mfa/reset:
    post:
      description: |-
        Clears the target's TOTP secret and recovery codes — the target
        must enroll again at next login and loses every active session.
      parameters:
      - description: Admin user UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Reset another platform admin's MFA
      tags:
      - platform
...
  /platform/mfa/confirm:
    post:
      consumes:
      - application/json
      description: |-
        Verifies the code against the secret from PlatformMFASetupHandler,
        requires password re-confirmation, persists the secret, and
        returns one-time recovery codes. Re-enrollment goes through an
        MFA reset by another admin instead.
      parameters:
      - description: Secret, code, password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.platformMFAConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm platform admin TOTP enrollment
      tags:
      - platform
  /platform/mfa/enroll:
    post:
      consumes:
      - application/json
      description: |-
        Public. Completes the mandatory setup started by a login that
        returned mfa_setup_required: verifies the enrollment token and
        TOTP code, stores the secret, and returns one-time recovery codes.
      parameters:
      - description: Enrollment token, secret, code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.platformMFAEnrollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/server.apiErrorResponse'
      summary: Finish platform admin TOTP enrollment with a login-time enrollment
        token
      tags:
      - platform
  /platform/mfa/setup:
    post:
      description: |-
        Writes nothing to the database — the secret only persists once
        PlatformMFAConfirmHandler verifies it.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
      security:
      - BearerAuth: []
      summary: Generate a new platform admin TOTP secret
      tags:
      - platform
...
swagger: "2.0"